Refactored all maps files.
