+++ app/src/main/java/no/hiof/mariusca/stitur/ui/screen/map/utils/PermissionBox.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package no.hiof.mariusca.stitur.ui.screen.map
+package no.hiof.mariusca.stitur.ui.screen.map.utils
 
 import android.content.Intent
 import android.net.Uri
